More randomness in particle values, static_cast<float> on the rand() method

+++ src/particles.cpp
@@ -12,8 +12,8 @@ Particles::Particles(int nb_types_particles, int nb_particles_per_type){
     srand(time(NULL));
 
     for(int i=0;i<nb_types_particles;i++){
-        float randSpeed = rand() % 200 + 50;
-        float randRadius = rand() % 5 + 0.5;
+        float randSpeed = static_cast<float>(rand() % 200 + 50);
+        float randRadius = static_cast<float>(rand() % 5 + 0.5);
         int randRed = rand() % 255;
         int randGreen = rand() % 255;
         int randBlue = rand() % 255;
@@ -24,8 +24,8 @@ Particles::Particles(int nb_types_particles, int nb_particles_per_type){
         float randWeight = 10.0 * normalizedRadius * normalizedRadius + 1; // Weight proportional to the square of the normalized radius
 
         for(int j=0;j<nb_particles_per_type;j++){
-            float randX = rand() % WINDOW_WIDTH + randRadius;
-            float randY = rand() % WINDOW_HEIGHT + randRadius;
+            float randX = static_cast<float>(rand() % WINDOW_WIDTH + randRadius);
+            float randY = static_cast<float>(rand() % WINDOW_HEIGHT + randRadius);
             float randDirection = static_cast<float>(rand()) / static_cast<float>(RAND_MAX) * 2.0 * M_PI;
 
             Particle particle(randX, randY, randSpeed, randDirection, randRadius, randRed, randGreen, randBlue, randWeight);
@@ -40,8 +40,8 @@ Particles::Particles(int nb_types_particles){
     for(int i=0;i<nb_types_particles;i++){
         int nb_particles = rand() % max_particles_per_type + 1;
 
-        float randSpeed = rand() % 200 + 50;
-        float randRadius = rand() % 5 + 0.5;
+        float randSpeed = static_cast<float>(rand() % 200 + 50);
+        float randRadius = static_cast<float>(rand() % 5 + 0.5);
         int randRed = rand() % 255;
         int randGreen = rand() % 255;
         int randBlue = rand() % 255;
@@ -52,8 +52,8 @@ Particles::Particles(int nb_types_particles){
         float randWeight = 10.0 * normalizedRadius * normalizedRadius + 1; // Weight proportional to the square of the normalized radius
 
         for(int j=0;j<max_particles_per_type;j++){
-            float randX = rand() % WINDOW_WIDTH + randRadius;
-            float randY = rand() % WINDOW_HEIGHT + randRadius;
+            float randX = static_cast<float>(rand() % WINDOW_WIDTH + randRadius);
+            float randY = static_cast<float>(rand() % WINDOW_HEIGHT + randRadius);
             float randDirection = static_cast<float>(rand()) / static_cast<float>(RAND_MAX) * 2.0 * M_PI;
 
             Particle particle(randX, randY, randSpeed, randDirection, randRadius, randRed, randGreen, randBlue, randWeight);
